Code cleanup.

+++ fest-assert/src/test/java/org/fest/assertions/ByteAssert_satisfies_Test.java
@@ -24,7 +24,7 @@ package org.fest.assertions;
 public class ByteAssert_satisfies_Test extends GenericAssert_satisfies_TestCase<Byte> {
 
   protected ByteAssert assertObject() {
-    return new ByteAssert((byte)6);
+    return new ByteAssert(6);
   }
 
   protected ByteAssert assertObjectWithNullTarget() {
